+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_android/lib/src/kevin_flutter_accounts_android.dart
@@ -23,7 +23,7 @@ class KevinFlutterAccountsAndroid
       KevinFlutterAccountsMethods.setAccountsConfiguration,
       {
         'callbackUrl': configuration.callbackUrl.android,
-        'showUnsupportedBanks': configuration.showUnsupportedBanks
+        'showUnsupportedBanks': configuration.showUnsupportedBanks,
       },
     );
   }