+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_android/test/kevin_flutter_accounts_android_test.dart
@@ -15,7 +15,8 @@ void main() {
   final platform = KevinFlutterAccountsAndroid();
 
   void _setMethodCallReturnData({dynamic Function()? data}) {
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, (MethodCall methodCall) {
       log.add(methodCall);
       return data?.call();
     });
@@ -48,9 +49,9 @@ void main() {
         'setAccountsConfiguration',
         arguments: <String, dynamic>{
           'callbackUrl': 'android',
-          'showUnsupportedBanks': true
+          'showUnsupportedBanks': true,
         },
-      )
+      ),
     ]);
   });
 
@@ -87,7 +88,7 @@ void main() {
           'skipBankSelection': false,
           'accountLinkingType': 'bank',
         },
-      )
+      ),
     ]);
   });
 
@@ -114,7 +115,7 @@ void main() {
           'skipBankSelection': false,
           'accountLinkingType': 'bank',
         },
-      )
+      ),
     ]);
   });
 
@@ -141,7 +142,7 @@ void main() {
           'skipBankSelection': false,
           'accountLinkingType': 'bank',
         },
-      )
+      ),
     ]);
   });
 
@@ -168,7 +169,7 @@ void main() {
           'skipBankSelection': false,
           'accountLinkingType': 'bank',
         },
-      )
+      ),
     ]);
   });
 
@@ -193,7 +194,7 @@ void main() {
           'skipBankSelection': false,
           'accountLinkingType': 'bank',
         },
-      )
+      ),
     ]);
   });
 
@@ -205,7 +206,7 @@ void main() {
       isMethodCall(
         'getCallbackUrl',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -220,7 +221,7 @@ void main() {
       isMethodCall(
         'getCallbackUrl',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -232,7 +233,7 @@ void main() {
       isMethodCall(
         'isShowUnsupportedBanks',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -247,7 +248,7 @@ void main() {
       isMethodCall(
         'isShowUnsupportedBanks',
         arguments: null,
-      )
+      ),
     ]);
   });
 }