+++ packages/kevin_flutter_core/kevin_flutter_core_ios/test/kevin_flutter_core_ios_test.dart
@@ -13,10 +13,14 @@ void main() {
   final platform = KevinFlutterCoreIos();
 
   void _setMethodCallReturnData({dynamic Function()? data}) {
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      log.add(methodCall);
-      return data?.call();
-    });
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(
+      channel,
+      (MethodCall methodCall) async {
+        log.add(methodCall);
+        return data?.call();
+      },
+    );
   }
 
   setUp(() {
@@ -41,7 +45,7 @@ void main() {
       isMethodCall(
         'setLocale',
         arguments: <String, String>{'languageCode': 'en'},
-      )
+      ),
     ]);
   });
 
@@ -81,8 +85,12 @@ void main() {
       ),
       gridTableStyle: const KevinGridTableStyle(
         cellBackgroundColor: Colors.black,
-        cellSelectedBackgroundColor: Colors.blue,
         cellCornerRadius: 16,
+        cellBorderWidth: 2,
+        cellBorderColor: Colors.blue,
+        cellSelectedBackgroundColor: Colors.blue,
+        cellSelectedBorderWidth: 2,
+        cellSelectedBorderColor: Colors.blue,
       ),
       listTableStyle: KevinListTableStyle(
         cornerRadius: 10,
@@ -153,7 +161,7 @@ void main() {
       isMethodCall(
         'setTheme',
         arguments: <String, dynamic>{'theme': themeIos.toMap()},
-      )
+      ),
     ]);
   });
 
@@ -168,7 +176,7 @@ void main() {
       isMethodCall(
         'setSandbox',
         arguments: <String, bool>{'sandbox': true},
-      )
+      ),
     ]);
   });
 
@@ -178,7 +186,7 @@ void main() {
       isMethodCall(
         'setDeepLinkingEnabled',
         arguments: <String, bool>{'deepLinkingEnabled': true},
-      )
+      ),
     ]);
   });
 
@@ -191,7 +199,7 @@ void main() {
       isMethodCall(
         'getLocale',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -202,7 +210,7 @@ void main() {
       isMethodCall(
         'getLocale',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -215,7 +223,7 @@ void main() {
       isMethodCall(
         'isSandbox',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -230,7 +238,7 @@ void main() {
       isMethodCall(
         'isSandbox',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -243,7 +251,7 @@ void main() {
       isMethodCall(
         'isDeepLinkingEnabled',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -258,7 +266,7 @@ void main() {
       isMethodCall(
         'isDeepLinkingEnabled',
         arguments: null,
-      )
+      ),
     ]);
   });
 }