+++ packages/kevin_flutter_core/kevin_flutter_core_platform_interface/lib/src/model/theme/ios/kevin_grid_table_style.dart
@@ -2,20 +2,35 @@ import 'package:flutter/painting.dart';
 
 class KevinGridTableStyle {
   final Color? cellBackgroundColor;
-  final Color? cellSelectedBackgroundColor;
+
   final double? cellCornerRadius;
+  final double? cellBorderWidth;
+  final Color? cellBorderColor;
+
+  final Color? cellSelectedBackgroundColor;
+
+  final double? cellSelectedBorderWidth;
+  final Color? cellSelectedBorderColor;
 
   const KevinGridTableStyle({
     this.cellBackgroundColor,
-    this.cellSelectedBackgroundColor,
     this.cellCornerRadius,
+    this.cellBorderWidth,
+    this.cellBorderColor,
+    this.cellSelectedBackgroundColor,
+    this.cellSelectedBorderWidth,
+    this.cellSelectedBorderColor,
   });
 
   Map<String, dynamic> toMap() {
     return {
       'cellBackgroundColor': cellBackgroundColor?.value,
-      'cellSelectedBackgroundColor': cellSelectedBackgroundColor?.value,
       'cellCornerRadius': cellCornerRadius,
+      'cellBorderWidth': cellBorderWidth,
+      'cellBorderColor': cellBorderColor?.value,
+      'cellSelectedBackgroundColor': cellSelectedBackgroundColor?.value,
+      'cellSelectedBorderWidth': cellSelectedBorderWidth,
+      'cellSelectedBorderColor': cellSelectedBorderColor?.value,
     };
   }
 }