+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_platform_interface/test/kevin_flutter_method_channel_test.dart
@@ -21,10 +21,14 @@ void main() {
   final initialInstance = KevinFlutterPaymentsPlatformInterface.instance;
 
   void _setMethodCallReturnData({dynamic Function()? data}) {
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      log.add(methodCall);
-      return data?.call();
-    });
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(
+      channel,
+      (MethodCall methodCall) async {
+        log.add(methodCall);
+        return data?.call();
+      },
+    );
   }
 
   setUp(() {
@@ -70,7 +74,7 @@ void main() {
         arguments: <String, dynamic>{
           'callbackUrl': '',
         },
-      )
+      ),
     ]);
   });
 
@@ -100,7 +104,7 @@ void main() {
           'skipBankSelection': false,
           'skipAuthentication': false,
         },
-      )
+      ),
     ]);
   });
 
@@ -129,7 +133,7 @@ void main() {
           'skipBankSelection': false,
           'skipAuthentication': false,
         },
-      )
+      ),
     ]);
   });
 
@@ -158,7 +162,7 @@ void main() {
           'skipBankSelection': false,
           'skipAuthentication': false,
         },
-      )
+      ),
     ]);
   });
 
@@ -187,7 +191,7 @@ void main() {
           'skipBankSelection': false,
           'skipAuthentication': false,
         },
-      )
+      ),
     ]);
   });
 
@@ -214,7 +218,7 @@ void main() {
           'skipBankSelection': false,
           'skipAuthentication': false,
         },
-      )
+      ),
     ]);
   });
 
@@ -227,7 +231,7 @@ void main() {
       isMethodCall(
         'getCallbackUrl',
         arguments: null,
-      )
+      ),
     ]);
   });
 
@@ -242,7 +246,7 @@ void main() {
       isMethodCall(
         'getCallbackUrl',
         arguments: null,
-      )
+      ),
     ]);
   });
 }