+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_android/CHANGELOG.md
@@ -1,9 +1,18 @@
+## 1.1.0
+
+* Update Dart min version to 3.0.0
+* Upgrade:
+    - `eu.kevin.android:core:2.7.0`
+    - `eu.kevin.android:accounts:2.7.0`
+    - `kevin_flutter_core: 1.1.0`
+    - `kevin_flutter_accounts_platform_interface: 1.1.0`
+
 ## 1.0.3
 
 * Upgrade:
-  - `eu.kevin.android:core:2.6.1`
-  - `kevin_flutter_core: 1.0.6`
-  - `kevin_flutter_accounts_platform_interface: 1.0.5`
+    - `eu.kevin.android:core:2.6.1`
+    - `kevin_flutter_core: 1.0.6`
+    - `kevin_flutter_accounts_platform_interface: 1.0.5`
 
 ## 1.0.2