+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_ios/CHANGELOG.md
@@ -1,9 +1,17 @@
+## 1.1.0
+
+* Update Dart min version to 3.0.0
+* Upgrade:
+    - `kevin_ios: 2.3.0`
+    - `kevin_flutter_core: 1.1.0`
+    - `kevin_flutter_accounts_platform_interface: 1.1.0`
+
 ## 1.0.3
 
 * Upgrade:
-  - `kevin_ios: 2.2.9`
-  - `kevin_flutter_core: 1.0.6`
-  - `kevin_flutter_accounts_platform_interface: 1.0.5`
+    - `kevin_ios: 2.2.9`
+    - `kevin_flutter_core: 1.0.6`
+    - `kevin_flutter_accounts_platform_interface: 1.0.5`
 
 ## 1.0.2