+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_android/CHANGELOG.md
@@ -1,9 +1,18 @@
+## 1.1.0
+
+* Update Dart min version to 3.0.0
+* Upgrade:
+    - `eu.kevin.android:core:2.7.0`
+    - `eu.kevin.android:in-app-payments:2.7.0`
+    - `kevin_flutter_core: 1.1.0`
+    - `kevin_flutter_in_app_payments_platform_interface: 1.1.0`
+
 ## 1.0.3
 
 * Upgrade:
-  - `eu.kevin.android:core:2.6.1`
-  - `kevin_flutter_core: 1.0.6`
-  - `kevin_flutter_in_app_payments_platform_interface: 1.0.4`
+    - `eu.kevin.android:core:2.6.1`
+    - `kevin_flutter_core: 1.0.6`
+    - `kevin_flutter_in_app_payments_platform_interface: 1.0.4`
 
 ## 1.0.2